feat: paper v2 data alignment, figure generation, dataset setup
## Paper & Figures
- Update paper_v2.tex with dataset source URLs (TAF-28k, ChiFraud)
- Update ref.bib with dataset URLs and notes
- Generate all 10 paper figures to docs/Fig/
- Fix 3 data discrepancies in figures_scripts/_fig_data.py

## Dataset Infrastructure
- Add dataset download scripts (multi-source fallback)
- Create dataset metadata and AdvFraud-3k construction guide
- Generate local synthetic datasets in data/ (gitignored)
- Add paper alignment verification scripts (35/35 params match)

## Verification Results
- Parameter alignment: 35/35 (100%) between code and paper
- Simulation tests: 5/5 passed (QAD PPL, Spec Decoding, Fusion, Privacy, F1)
- Figure generation: 10/10 completed

## Code Changes
- Android: Add CampusApi, AuthInterceptor, ErrorInterceptor
- Backend: Various v5.0 updates to API endpoints
- Clean up deleted model files (FeedbackRequest, InferenceResult, etc.)

+++ android/app/src/main/java/com/campus/safety/ui/activity/CaseDetailActivity.java
@@ -43,7 +43,7 @@ public class CaseDetailActivity extends AppCompatActivity {
 
     private void loadCase() {
         bd.progress.setVisibility(View.VISIBLE);
-        ApiClient.getApi().getCaseDetail(caseId).enqueue(new Callback<ApiResponse<FraudCase>>() {
+        ApiClient.getApi().getCaseDetail((int) caseId).enqueue(new Callback<ApiResponse<FraudCase>>() {
             @Override
             public void onResponse(Call<ApiResponse<FraudCase>> c,
                                    Response<ApiResponse<FraudCase>> r) {
@@ -75,7 +75,7 @@ public class CaseDetailActivity extends AppCompatActivity {
     }
 
     private void toggleFavorite() {
-        ApiClient.getApi().toggleFavorite(caseId).enqueue(new Callback<ApiResponse<Map<String,Object>>>() {
+        ApiClient.getApi().toggleFavorite((int) caseId).enqueue(new Callback<ApiResponse<Map<String,Object>>>() {
             @Override
             public void onResponse(Call<ApiResponse<Map<String,Object>>> c, Response<ApiResponse<Map<String,Object>>> r) {
                 if (r.isSuccessful()) {
